feat: support loading node native addons from the prebuilds directory (#198)
Extracted from #194

This is taking advantage of #185, which used a function in
node-gyp-build to compute the path. Some moons ago,
[node-gyp-build.js#L62-L75](https://github.com/prebuild/node-gyp-build/blob/v4.8.4/node-gyp-build.js#L62-L75)
was updated with a function to resolve native addons from the prebuilds
directory. All we have to do here is update `node-gyp-build`.

This enables assets relocation for packages like [bcrypt
v6.0.0](https://www.npmjs.com/package/bcrypt?activeTab=code)

I added some tests according to the platform used in CI. Hopefully, the
architecture and the fixtures align well.

---------

Co-authored-by: Steven <steven@ceriously.com>